+++ lib/docker_commander_browser.dart
@@ -1,0 +1,4 @@
+library docker_commander;
+
+export 'docker_commander.dart';
+export 'src/docker_commander_remote.dart';