+++ lib/docker_commander_vm.dart
@@ -1,0 +1,6 @@
+library docker_commander;
+
+export 'docker_commander.dart';
+export 'src/docker_commander_local.dart';
+export 'src/docker_commander_server.dart';
+export 'src/docker_commander_remote.dart';